Arena Playtest Levels Blocked Out
The tutorial level is fully blocked out and functional with enemies, fake walls, and arenas. It still needs the boss placeholder and a transition to the second level. The second level has spirits and save points set up, but still needs enemies in both the arena and the general level